% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CRTConjoint-package.R
\docType{package}
\name{CRTConjoint-package}
\alias{CRTConjoint}
\alias{CRTConjoint-package}
\title{CRTConjoint: Conditional Randomization Testing (CRT) Approach for Conjoint Analysis}
\description{
Computes p-value according to the CRT using the HierNet test statistic. For more details, see Ham, Imai, Janson (2022) "Using Machine Learning to Test Causal Hypotheses in Conjoint Analysis".
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/daewoongham97/CRTConjoint}
  \item Report bugs at \url{https://github.com/daewoongham97/CRTConjoint/issues}
}

}
\keyword{internal}
